added limits to all input fields for update and insert

+++ project2-audit-triggers.sql
@@ -8,7 +8,7 @@ DROP TABLE IF EXISTS violations_audit;
 CREATE TABLE violations_audit (
 	audit_id INT AUTO_INCREMENT,
     summons_number BIGINT,
-    violation_code TINYINT, 
+    violation_code INT, 
     violation_legal_code CHAR(1), 
     date_first_observed VARCHAR(10), 
     violation_description MEDIUMTEXT,
